refactor(download): applied the new folder structure to the download section related components
