+++ src/components/Form/Form.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import QuerySection from '@/sections/query/QuerySection';
 import UsageSection from '@/sections/usage/UsageSection';
-import DownloadSection from '../DownloadSection/DownloadSection';
+import DownloadSection from '@/sections/download/DownloadSection';
 
 export default function Form () {
   return (
